ElasticSearchDestination-v2: native MessageProcessors moved
to messageprocessor.esnative
